Multiple T's defined caused problems.
Need to clarify when to use T,Tmax,t,tmax
would propose that capitals are for symbols and lower case for
numerical.

Additional problem is that stress tensor is also T, makes it confusing.
Maybe this should be T_cauchy?

View the git diff to see all changes... there were many small ones.

Notably I changed the Rt and Bt to be functions so that we don't end up
with something hardcoded by a variable assignment.

Note that the charts look wrong at the end... need to compare to someone
else.  I had them looking more like the professor's at one point, but I
think the code is more correct now.

Problem (6) had an error because the matrices were in the wrong order
for the composition.